styled ofr grocery list items added. Scrool bar still needed
