+++ cookbook/src/Styles/less/index.less
@@ -30,6 +30,8 @@ html {
 
 @import "grocery-list.less";
 
+@import 'grocery-item.less';
+
 @import "recipes.less";
 
 @import "recipe-card.less";
